+++ src/dynamic-pages/analyze/charts/chart.tsx
@@ -5,12 +5,12 @@ import { useTheme } from '@mui/material/styles';
 import useMediaQuery from '@mui/material/useMediaQuery';
 import EChartsReact from 'echarts-for-react';
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import CommonChartContext, { CommonChartShareInfo } from '../components/CommonChart/context';
-import DebugDialog from '../components/DebugDialog/DebugDialog';
-import ShareDialog from '../components/ShareDialog';
+import CommonChartContext, { CommonChartShareInfo } from '../../../components/CommonChart/context';
+import DebugDialog from '../../../components/DebugDialog/DebugDialog';
+import ShareDialog from '../../../components/ShareDialog';
 import {AnalyzeChartContextProps, AnalyzeContextProps, useAnalyzeChartContext, useAnalyzeContext} from './context';
 import {EChartsOption} from 'echarts';
-import ECharts, { EChartsContext, EChartsProps } from '../components/ECharts';
+import ECharts, { EChartsContext, EChartsProps } from '../../../components/ECharts';
 import { DangerousCtx, dangerousSetCtx } from './options/_danger';
 import useDimensions from 'react-cool-dimensions';
 import { debounce } from '@mui/material';
