+++ src/dynamic-pages/analyze/charts/companies/index.ts
