+++ src/dynamic-pages/analyze/charts/line/index.tsx
