+++ src/dynamic-pages/analyze/charts/list/index.tsx
