+++ src/dynamic-pages/analyze/charts/list/styled.tsx
