+++ src/dynamic-pages/analyze/charts/options/datazoom.ts
