+++ src/dynamic-pages/analyze/charts/options/legend.ts
