+++ src/dynamic-pages/analyze/charts/options/series.ts
