+++ src/dynamic-pages/analyze/charts/options/utils/data.ts
