+++ src/dynamic-pages/analyze/charts/options/visual-map.ts
