+++ src/dynamic-pages/analyze/charts/summary/Summary.tsx
